+++ src/main/java/ph/com/paraiso/dao/impl/UserDaoImpl.java
@@ -33,7 +33,7 @@ public class UserDaoImpl implements UserDao{
 	}
 
 	@Override
-	public User getUserById(Long userid) {
+	public User getUserById(Integer userid) {
 		return userRepo.findById(userid).get();
 	}
 
@@ -43,7 +43,7 @@ public class UserDaoImpl implements UserDao{
 	}
 
 	@Override
-	public void deleteById(Long userid) {
+	public void deleteById(Integer userid) {
 		userRepo.deleteById(userid);
 		
 	}
